Moved the default context (DevContext which handles request parameters) from lib/web to app/filters as params.go
Refactored context chaining to remove doubly linked list [as we use simple iteration anyways]
as well as passing the chain list into TestContext.

This way you can test that the context-chain you belong to supports features you need to
service a request.

This allows, for e.g, the FlashContext to piggyback on the SessionContext for session storage.