Minor fix b/c I don't like setting rowspan="0"
It seems to work in Chrome but documented behavior means that the cell could
legally expand to the end of the table; this is not what we want.

If NumEpisodes() is 0 it _really_ means that this is a complete series in a single .torrent,
so we return "1" which gives us room for a single line item.